+++ sample/src/com/squareup/timessquare/sample/SampleTimesSquareActivity.java
@@ -10,11 +10,17 @@ import android.view.View;
 import android.view.View.OnClickListener;
 import android.widget.Button;
 import android.widget.Toast;
+import com.squareup.timessquare.CalendarCellDecorator;
 import com.squareup.timessquare.CalendarPickerView;
 import com.squareup.timessquare.CalendarPickerView.SelectionMode;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Calendar;
+import java.util.Collections;
 import java.util.Date;
+import java.util.LinkedHashSet;
+import java.util.Locale;
+import java.util.Set;
 
 import static android.widget.Toast.LENGTH_SHORT;
 
@@ -23,9 +29,9 @@ public class SampleTimesSquareActivity extends Activity {
   private CalendarPickerView calendar;
   private AlertDialog theDialog;
   private CalendarPickerView dialogView;
+  private final Set<Button> modeButtons = new LinkedHashSet<Button>();
 
-  @Override
-  protected void onCreate(Bundle savedInstanceState) {
+  @Override protected void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
     setContentView(R.layout.sample_calendar_picker);
 
@@ -40,20 +46,26 @@ public class SampleTimesSquareActivity extends Activity {
         .inMode(SelectionMode.SINGLE) //
         .withSelectedDate(new Date());
 
+    initButtonListeners(nextYear, lastYear);
+  }
+
+  private void initButtonListeners(final Calendar nextYear, final Calendar lastYear) {
     final Button single = (Button) findViewById(R.id.button_single);
     final Button multi = (Button) findViewById(R.id.button_multi);
     final Button range = (Button) findViewById(R.id.button_range);
     final Button displayOnly = (Button) findViewById(R.id.button_display_only);
     final Button dialog = (Button) findViewById(R.id.button_dialog);
     final Button customized = (Button) findViewById(R.id.button_customized);
-    single.setOnClickListener(new OnClickListener() {
-      @Override
-      public void onClick(View v) {
-        single.setEnabled(false);
-        multi.setEnabled(true);
-        range.setEnabled(true);
-        displayOnly.setEnabled(true);
+    final Button decorator = (Button) findViewById(R.id.button_decorator);
+    final Button rtl = (Button) findViewById(R.id.button_rtl);
 
+    modeButtons.addAll(Arrays.asList(single, multi, range, displayOnly, decorator));
+
+    single.setOnClickListener(new OnClickListener() {
+      @Override public void onClick(View v) {
+        setButtonsEnabled(single);
+
+        calendar.setDecorators(Collections.<CalendarCellDecorator>emptyList());
         calendar.init(lastYear.getTime(), nextYear.getTime()) //
             .inMode(SelectionMode.SINGLE) //
             .withSelectedDate(new Date());
@@ -61,12 +73,8 @@ public class SampleTimesSquareActivity extends Activity {
     });
 
     multi.setOnClickListener(new OnClickListener() {
-      @Override
-      public void onClick(View v) {
-        single.setEnabled(true);
-        multi.setEnabled(false);
-        range.setEnabled(true);
-        displayOnly.setEnabled(true);
+      @Override public void onClick(View v) {
+        setButtonsEnabled(multi);
 
         Calendar today = Calendar.getInstance();
         ArrayList<Date> dates = new ArrayList<Date>();
@@ -74,6 +82,7 @@ public class SampleTimesSquareActivity extends Activity {
           today.add(Calendar.DAY_OF_MONTH, 3);
           dates.add(today.getTime());
         }
+        calendar.setDecorators(Collections.<CalendarCellDecorator>emptyList());
         calendar.init(new Date(), nextYear.getTime()) //
             .inMode(SelectionMode.MULTIPLE) //
             .withSelectedDates(dates);
@@ -81,12 +90,8 @@ public class SampleTimesSquareActivity extends Activity {
     });
 
     range.setOnClickListener(new OnClickListener() {
-      @Override
-      public void onClick(View v) {
-        single.setEnabled(true);
-        multi.setEnabled(true);
-        range.setEnabled(false);
-        displayOnly.setEnabled(true);
+      @Override public void onClick(View v) {
+        setButtonsEnabled(range);
 
         Calendar today = Calendar.getInstance();
         ArrayList<Date> dates = new ArrayList<Date>();
@@ -94,6 +99,7 @@ public class SampleTimesSquareActivity extends Activity {
         dates.add(today.getTime());
         today.add(Calendar.DATE, 5);
         dates.add(today.getTime());
+        calendar.setDecorators(Collections.<CalendarCellDecorator>emptyList());
         calendar.init(new Date(), nextYear.getTime()) //
             .inMode(SelectionMode.RANGE) //
             .withSelectedDates(dates);
@@ -101,79 +107,86 @@ public class SampleTimesSquareActivity extends Activity {
     });
 
     displayOnly.setOnClickListener(new OnClickListener() {
-      @Override
-      public void onClick(View v) {
-        single.setEnabled(true);
-        multi.setEnabled(true);
-        range.setEnabled(true);
-        displayOnly.setEnabled(false);
+      @Override public void onClick(View v) {
+        setButtonsEnabled(displayOnly);
 
+        calendar.setDecorators(Collections.<CalendarCellDecorator>emptyList());
         calendar.init(new Date(), nextYear.getTime()) //
             .inMode(SelectionMode.SINGLE) //
-            .withSelectedDate(new Date())
+            .withSelectedDate(new Date()) //
             .displayOnly();
       }
     });
 
     dialog.setOnClickListener(new OnClickListener() {
       @Override public void onClick(View view) {
-        dialogView = (CalendarPickerView) getLayoutInflater().inflate(R.layout.dialog, null, false);
+        String title = "I'm a dialog!";
+        showCalendarInDialog(title, R.layout.dialog);
         dialogView.init(lastYear.getTime(), nextYear.getTime()) //
             .withSelectedDate(new Date());
-        theDialog =
-            new AlertDialog.Builder(SampleTimesSquareActivity.this).setTitle("I'm a dialog!")
-                .setView(dialogView)
-                .setNeutralButton("Dismiss", new DialogInterface.OnClickListener() {
-                  @Override public void onClick(DialogInterface dialogInterface, int i) {
-                    dialogInterface.dismiss();
-                  }
-                })
-                .create();
-        theDialog.setOnShowListener(new DialogInterface.OnShowListener() {
-          @Override
-          public void onShow(DialogInterface dialogInterface) {
-            Log.d(TAG, "onShow: fix the dimens!");
-            dialogView.fixDialogDimens();
-          }
-        });
-        theDialog.show();
       }
     });
 
     customized.setOnClickListener(new OnClickListener() {
-      @Override
-      public void onClick(View view) {
-        dialogView = (CalendarPickerView) getLayoutInflater() //
-            .inflate(R.layout.dialog_customized, null, false);
-        dialogView.init(lastYear.getTime(), nextYear.getTime()).withSelectedDate(new Date());
-        theDialog =
-            new AlertDialog.Builder(SampleTimesSquareActivity.this).setTitle("Pimp my calendar !")
-                .setView(dialogView)
-                .setNeutralButton("Dismiss", new DialogInterface.OnClickListener() {
-                  @Override
-                  public void onClick(DialogInterface dialogInterface, int i) {
-                    dialogInterface.dismiss();
-                  }
-                }).create();
-        theDialog.setOnShowListener(new DialogInterface.OnShowListener() {
-          @Override
-          public void onShow(DialogInterface dialogInterface) {
-            Log.d(TAG, "onShow: fix the dimens!");
-            dialogView.fixDialogDimens();
-          }
-        });
-        theDialog.show();
+      @Override public void onClick(View view) {
+        showCalendarInDialog("Pimp my calendar!", R.layout.dialog_customized);
+        dialogView.init(lastYear.getTime(), nextYear.getTime())
+            .withSelectedDate(new Date());
+      }
+    });
+
+    decorator.setOnClickListener(new OnClickListener() {
+      @Override public void onClick(View v) {
+        setButtonsEnabled(decorator);
+
+        calendar.setDecorators(Arrays.<CalendarCellDecorator>asList(new SampleDecorator()));
+        calendar.init(lastYear.getTime(), nextYear.getTime()) //
+            .inMode(SelectionMode.SINGLE) //
+            .withSelectedDate(new Date());
+      }
+    });
+
+    rtl.setOnClickListener(new OnClickListener() {
+      @Override public void onClick(View view) {
+        showCalendarInDialog("I'm right-to-left!", R.layout.dialog);
+        dialogView.init(lastYear.getTime(), nextYear.getTime(), new Locale("iw", "IL")) //
+            .withSelectedDate(new Date());
       }
     });
 
     findViewById(R.id.done_button).setOnClickListener(new OnClickListener() {
-      @Override
-      public void onClick(View view) {
+      @Override public void onClick(View view) {
         Log.d(TAG, "Selected time in millis: " + calendar.getSelectedDate().getTime());
         String toast = "Selected: " + calendar.getSelectedDate().getTime();
         Toast.makeText(SampleTimesSquareActivity.this, toast, LENGTH_SHORT).show();
       }
     });
+  }
+
+  private void showCalendarInDialog(String title, int layoutResId) {
+    dialogView = (CalendarPickerView) getLayoutInflater().inflate(layoutResId, null, false);
+    theDialog = new AlertDialog.Builder(this) //
+        .setTitle(title)
+        .setView(dialogView)
+        .setNeutralButton("Dismiss", new DialogInterface.OnClickListener() {
+          @Override public void onClick(DialogInterface dialogInterface, int i) {
+            dialogInterface.dismiss();
+          }
+        })
+        .create();
+    theDialog.setOnShowListener(new DialogInterface.OnShowListener() {
+      @Override public void onShow(DialogInterface dialogInterface) {
+        Log.d(TAG, "onShow: fix the dimens!");
+        dialogView.fixDialogDimens();
+      }
+    });
+    theDialog.show();
+  }
+
+  private void setButtonsEnabled(Button currentButton) {
+    for (Button modeButton : modeButtons) {
+      modeButton.setEnabled(modeButton != currentButton);
+    }
   }
 
   @Override public void onConfigurationChanged(Configuration newConfig) {